# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TripSync API",
    "version": "1.0.0",
    "description": "API endpoints for TripSync - Trip Expense Splitter & Travel Planner backend."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "CreateTripRequest": {
        "type": "object",
        "required": ["title", "destination", "start_date", "end_date"],
        "properties": {
          "title": { "type": "string", "minLength": 2, "maxLength": 100 },
          "description": { "type": "string", "maxLength": 1000 },
          "destination": { "type": "string", "maxLength": 200 },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "currency": { "type": "string", "maxLength": 3, "default": "INR" }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/users/me": {
      "get": {
        "summary": "Get logged-in user profile",
        "responses": {
          "200": { "description": "Profile data retrieved" }
        }
      },
      "patch": {
        "summary": "Update logged-in user profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": { "type": "string" },
                  "username": { "type": "string" },
                  "bio": { "type": "string" },
                  "profile_photo_url": { "type": "string" },
                  "profile_photo_public_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated successfully" }
        }
      }
    },
    "/users/search": {
      "get": {
        "summary": "Search users for trip invitation",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Query term" }
        ],
        "responses": {
          "200": { "description": "Search results returned" }
        }
      }
    },
    "/trips": {
      "post": {
        "summary": "Create a new trip",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequest"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Trip created" }
        }
      },
      "get": {
        "summary": "List all trips user is part of",
        "responses": {
          "200": { "description": "List of trips" }
        }
      }
    },
    "/trips/{trip_id}": {
      "get": {
        "summary": "Get trip details",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Trip details" }
        }
      },
      "patch": {
        "summary": "Update trip details",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "destination": { "type": "string" },
                  "start_date": { "type": "string", "format": "date" },
                  "end_date": { "type": "string", "format": "date" },
                  "currency": { "type": "string" },
                  "status": { "type": "string", "enum": ["planning", "ongoing", "completed", "cancelled"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Trip updated" }
        }
      },
      "delete": {
        "summary": "Delete a trip",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Trip deleted" }
        }
      }
    },
    "/trips/{trip_id}/join": {
      "post": {
        "summary": "Join a trip by ID",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Joined successfully" }
        }
      }
    },
    "/trips/join-by-code": {
      "post": {
        "summary": "Join a trip using an invite code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["invite_code"],
                "properties": {
                  "invite_code": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Joined successfully" }
        }
      }
    },
    "/trips/{trip_id}/members": {
      "get": {
        "summary": "Get all members of a trip",
        "parameters": [
          { "name": "trip_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of members" }
        }
      }
    },
...
  }
}